requirements.txt
streamlit
faiss-cpu
sentence-transformers
mysql-connector-python
pypdf
numpy